Connection error fixed:
There was an error that prevented to connecting to a saved network
successfullly.

This error isn't asociated to the wifi script as such, but to an 'error' from
the latest versions of nmcli tool (see https://bbs.archlinux.org/viewtopic.php?id=300321 for more info).

For this reason, it was necessary change the connection logic:

- A new function "_connect_to_saved_network(ssid)" was added to handle
  that specific case.
- The above function uses the command "nmcli connection up <ssid>" that was defined in the 'default.py'

NOTE: Due to the nature of the avobe command, no password is reuired to pick up the connection